requirements.txt
ultralytics
opencv-python
numpy
gspread
oauth2client
python-dotenv
requests